travis.yml: use the same image as Murdock.
It does not make sense to do the static tests in a diferent image than
the compilation tests.

This commit updates the travis config to use the riotbuild image (the same
as murdock). This should fix the issues with old doxygen versions reporting
issues that are not, and also simplify the CI maintainance.